requirements.txt
asgiref==3.8.1
astroid==3.3.10
colorama==0.4.6
dill==0.4.0
Django==4.2.21
djangorestframework==3.16.0
drf-yasg==1.21.10
inflection==0.5.1
isort==6.0.1
mccabe==0.7.0
packaging==25.0
platformdirs==4.3.8
pylint==3.3.7
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
tomlkit==0.13.2
tzdata==2025.2
uritemplate==4.1.1
